Refactor upgrades system
This refactoring changes all upgrade indexing to use a hashmap, as well as
de-hardcoding the upgradeinfo array, allowing you to add external upgrades.

Relevant headers have been split into an MIT licensed API, which for now
does not include the player structure due to... complications.

Also missing is the ability to add callbacks or directly giving an array
of upgradeinfo_t to the mod, but those will come later.