Change monster allocation method
Previously, this was done using world arrays (__hub_arr,) but
unfortunately this actually does not work as expected. It only
deallocates when moving between hubs, not regular Doom maps.
So, instead, it is now allocated to a global array space and
the data is recycled when the user changes world areas.